+++ src/com/MPIDirectCommunication.hpp
@@ -54,7 +54,7 @@ public:
    *
    * Precondition: a connection to the remote participant has been setup.
    */
-  virtual int getRemoteCommunicatorSize();
+  virtual size_t getRemoteCommunicatorSize();
 
   /**
    * @brief See precice::com::Communication::acceptConnection().
